Module mem6::fncallermod

fncallermod

Functions

blink_or_not_group_id
if there is already a group_id don't blink

call_function_node
html_templating functions that return a Node

call_function_string
html_templating functions that return a String

call_listener
html_templating functions for listeners get a clone of the VdomWeak

game_type_left_onclick
left arrow button

game_type_right_onclick
the arrow to the right

get_input_value
get value form input html element by id

group_id_joined
return the text for html template replace empty if there is no group_id

open_new_local_page
fn open new local page with # window.location.set_hash

open_new_tab
fn open new tab

svg_qrcode_to_node
qrcode svg