Struct mem6::gamedatamod::GameData

pub struct GameData {
    pub my_ws_uid: usize,
    pub my_nickname: String,
    pub my_player_number: usize,
    pub ws: WebSocket,
    pub players: Vec<Player>,
    pub players_ws_uid: String,
    pub game_status: GameStatus,
    pub card_grid_data: Vec<Card>,
    pub card_index_of_first_click: usize,
    pub card_index_of_second_click: usize,
    pub game_name: String,
    pub player_turn: usize,
    pub content_folders: Vec<String>,
    pub games_metadata: Option<GamesMetadata>,
    pub game_config: Option<GameConfig>,
    pub error_text: String,
    pub href: String,
    pub href_hash: String,
    pub is_reconnect: bool,
    pub msgs_waiting_ack: Vec<MsgInQueue>,
    pub show_debug_info: bool,
}

game data

Fields
my_ws_uid: usize

my ws client instance unique id. To not listen the echo to yourself.

my_nickname: String

my nickname

my_player_number: usize

What player am I

ws: WebSocket

web socket. used it to send message onclick.

players: Vec<Player>

players data as vector of player struct

players_ws_uid: String

the json string for the ws server to send msgs to other players only

game_status: GameStatus

game status: StatusStartPage,Player1,Player2

card_grid_data: Vec<Card>

vector of cards

card_index_of_first_click: usize

card index of first click

card_index_of_second_click: usize

card index of second click

game_name: String

content folder name

player_turn: usize

whose turn is now: player 1,2,3,...

content_folders: Vec<String>

content folders vector

games_metadata: Option<GamesMetadata>

games meta data

game_config: Option<GameConfig>

game_configs

error_text: String

error text

href: String

href

href_hash: String

href hash the local page #

is_reconnect: bool

is reconnect

msgs_waiting_ack: Vec<MsgInQueue>

vector of msgs waiting for ack. If the 3 sec timeout passes it resends the same msg.

show_debug_info: bool

show debug info on the smartphone screen

Methods

impl GameData

pub fn prepare_random_data(&mut self)

prepare new random data

pub fn prepare_for_empty() -> Vec<Card>

associated function: before join, there are not random numbers, just default cards.

pub fn new(ws: WebSocket, my_ws_uid: usize) -> Self

constructor of game data