Module mem6::htmltemplatemod

htmltemplatemod
html templating for dodrio

Enums

HtmlOrSvg
Svg elements are different because they have a namespace

Functions

decode_5_xml_control_characters
decode 5 xml control characters : " ' & < >
https://www.liquid-technologies.com/XML/EscapingData.aspx I will ignore all html entities, to keep things simple, because all others characters can be written as utf-8 characters. https://www.tutorialspoint.com/html5/html5_entities.htm

empty_div
get en empty div node

fill_element_builder
Recursive function to fill the tree with a node.
Moves & Returns ElementBuilder or error.
I must move ElementBuilder because its methods are all move.
It makes the code less readable. It is only good for chaining and type changing.

get_root_element
get root element Node.
I wanted to use dodrio::Node, but it has only private methods.
I must use dodrio element_builder.